vision_mcp/pyproject.toml
[project]
name = "vision-mcp"
version = "0.1.0"
description = "GLM-4.6V Vision Model MCP Server"
authors = [{name = "algo-agent"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.2.0",
    "pillow>=10.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
vision-mcp = "vision_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
